Shop Love Sprint 15 bugs + revisions
This ticket will be ongoing for bugs and small revisions

Add MISLA tile to the carousel

When you click the "reservation" button for PIPS it leads to Freeway Easy.

When u click the website link button for aqua equity. It doesn’t go to their website like it should.

pls change the carousel scroll speed to be faster when user presses the arrow button. This is our 3rd request.